combined patient tests
