+++ tests/patient.test.js
@@ -1,4 +1,3 @@
-
 import chai from 'chai'
 const expect = chai.expect
 import chalk from 'chalk'
@@ -6,6 +5,12 @@ import bcrypt from 'bcrypt'
 
 import db from '../db'
 import Patient from '../db/models/patient'
+
+import app from '../server/app';
+import Promise from 'bluebird';
+import supertest from 'supertest';
+
+const agent = supertest.agent(app);
 
 describe('Patient', function () {
 
@@ -86,6 +91,47 @@ describe('Patient', function () {
           })
           .catch(err => console.log(err))
       })
+    })
+  })
+
+  describe.only('Patient Route', () => {
+    
+    describe('Patients', () => {
+
+      it('GET all patients', function () {
+        return agent
+        .get('/api/patient')
+        .expect(200)
+      });
+
+      it('POST one Patient', function (done) {
+        agent
+        .post('/api/patient')
+        .send({
+          'first_name': 'PatientTestF1',
+          'last_name': 'PatientTestL1',
+          'email': 'patient1@test.com',
+          'DOB': '11/18/15',
+          'gender': 'F'
+        })
+        .expect(201)
+        .end(function (err, res) {
+          console.log('sdfsdfsdfsd')
+          console.log('Error in Posting', err)
+
+          if (err) return done(err);
+          expect(res.body.first_name).to.equal('PatientTestF1');
+          expect(res.body.id).to.exist;
+          Patient.findById(res.body.id)
+            .then(function (b) {
+              expect(b).to.not.be.null;
+              expect(res.body).to.eql(toPlainObject(b));
+              done()
+            })
+            .catch(done);
+        })
+        .catch(err => console.log(err))
+      });
     })
   })
 })
